requirements.txt
Django==4.2.27
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Pillow
gunicorn==23.0.0


